+++ geotrek/core/sql/50_troncons_split.sql
@@ -145,19 +145,19 @@ BEGIN
         IF ST_DWITHIN(ST_STARTPOINT(NEW.geom), troncon.geom, 0)
         THEN
             intersections_on_current := array_append(intersections_on_current,
-                                                 ST_Line_Locate_Point(troncon.geom,
+                                                 ST_LineLocatePoint(troncon.geom,
                                                                       ST_CLOSESTPOINT(troncon.geom, ST_STARTPOINT(NEW.geom))));
         END IF;
 
         IF ST_DWITHIN(ST_ENDPOINT(NEW.geom), troncon.geom, 0)
         THEN
             intersections_on_current := array_append(intersections_on_current,
-                                                 ST_Line_Locate_Point(troncon.geom,
+                                                 ST_LineLocatePoint(troncon.geom,
                                                                       ST_CLOSESTPOINT(troncon.geom, ST_ENDPOINT(NEW.geom))));
 
         END IF;
         RAISE NOTICE 'EEE : %', array_to_string(intersections_on_current, ', ');
-        FOR fraction IN SELECT ST_Line_Locate_Point(troncon.geom, (ST_Dump(ST_Intersection(troncon.geom, NEW.geom))).geom)
+        FOR fraction IN SELECT ST_LineLocatePoint(troncon.geom, (ST_Dump(ST_Intersection(troncon.geom, NEW.geom))).geom)
         LOOP
             intersections_on_current := array_append(intersections_on_current, fraction);
         END LOOP;
@@ -187,7 +187,7 @@ BEGIN
                 a := intersections_on_new[i];
                 b := intersections_on_new[i+1];
 
-                segment := ST_Line_Substring(newgeom, a, b);
+                segment := ST_LineSubstring(newgeom, a, b);
 
                 IF coalesce(ST_Length(segment), 0) < 1 THEN
                      intersections_on_new[i+1] := a;
@@ -196,14 +196,14 @@ BEGIN
 
                 IF i = 1 THEN
                     -- First segment : shrink it !
-                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE nom = NEW.nom AND ST_Equals(geom, segment);
+                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE ST_Contains(ST_Buffer(segment,0.0001),geom);
                     IF t_count = 0 THEN
                         RAISE NOTICE 'New: Skrink %-% (%) to %', NEW.id, NEW.nom, ST_AsText(NEW.geom), ST_AsText(segment);
                         UPDATE l_t_troncon SET geom = segment WHERE id = NEW.id;
                     END IF;
                 ELSE
                     -- Next ones : create clones !
-                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE nom = NEW.nom AND ST_Equals(geom, segment);
+                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE ST_Contains(ST_Buffer(segment,0.0001),geom);
                     IF t_count = 0 THEN
                         RAISE NOTICE 'New: Create clone of %-% with geom %', NEW.id, NEW.nom, ST_AsText(segment);
                         INSERT INTO l_t_troncon (structure,
@@ -260,7 +260,7 @@ BEGIN
                 a := intersections_on_current[i];
                 b := intersections_on_current[i+1];
 
-                segment := ST_Line_Substring(troncon.geom, a, b);
+                segment := ST_LineSubstring(troncon.geom, a, b);
 
                 IF coalesce(ST_Length(segment), 0) < 1 THEN
                      intersections_on_new[i+1] := a;
@@ -276,7 +276,7 @@ BEGIN
                     END IF;
                 ELSE
                     -- Next ones : create clones !
-                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE ST_Equals(geom, segment);
+                    SELECT COUNT(*) INTO t_count FROM l_t_troncon WHERE ST_Contains(ST_Buffer(geom,0.0001),segment);
                     IF t_count = 0 THEN
                         RAISE NOTICE 'Current: Create clone of %-% (%) with geom %', troncon.id, troncon.nom, ST_AsText(troncon.geom), ST_AsText(segment);
                         INSERT INTO l_t_troncon (structure,
@@ -346,7 +346,7 @@ BEGIN
                         -- Special case : point topology at the end of path
                         IF b = 1 THEN
                             SELECT geom INTO t_geom FROM l_t_troncon WHERE id = troncon.id;
-                            fraction := ST_Line_Locate_Point(segment, ST_EndPoint(troncon.geom));
+                            fraction := ST_LineLocatePoint(segment, ST_EndPoint(troncon.geom));
                             INSERT INTO e_r_evenement_troncon (troncon, evenement, pk_debut, pk_fin)
                                 SELECT tid_clone, evenement, pk_debut, pk_fin
                                 FROM e_r_evenement_troncon et,
@@ -363,7 +363,7 @@ BEGIN
                         END IF;
                         -- Special case : point topology exactly where NEW path intersects
                         IF a > 0 THEN
-                            fraction := ST_Line_Locate_Point(NEW.geom, ST_Line_Interpolate_Point(troncon.geom, a));
+                            fraction := ST_LineLocatePoint(NEW.geom, ST_LineInterpolatePoint(troncon.geom, a));
                             INSERT INTO e_r_evenement_troncon (troncon, evenement, pk_debut, pk_fin, ordre)
                                 SELECT NEW.id, et.evenement, fraction, fraction, ordre
                                 FROM e_r_evenement_troncon et,
@@ -377,6 +377,7 @@ BEGIN
                                 RAISE NOTICE 'Duplicated % point topologies of %-% (%) on intersection by %-% (%) at [%]', t_count, troncon.id, troncon.nom, ST_AsText(troncon.geom), NEW.id, NEW.nom, ST_AsText(NEW.geom), a;
                             END IF;
                         END IF;
+
                     END IF;
                 END IF;
             END LOOP;